librsvg: fix thumbnailer path
librsvg hooks itself into gdk-pixbuf and then uses gdk-pixbuf-thumbnailer
as the thumbnailer, extending its supported MIME type list.

Unfortunately, librsvg assumes the thumbnailer will be located in the same
bindir as librsvg binaries would, which is not true on Nix-powered systems.

This commit corrects the bindir path of the thumbnailer to the gdk_pixbuf
derivation.