gnome3.gnome_control_center: fix thumbnailers path
Just like Nautilus (see #29970), GNOME Control Center also uses
gnome-desktop for generating thumbnails. In particular, it tries
to make a thumbnail from a file choosen as a profile picture, and
when it does not succeed, it will not allow that file to be chosen.
Of course, whithout a thumbnailer, it will always fail.

https://github.com/GNOME/gnome-control-center/blob/43129a1cfd30374c3ff45c54eae9ee5cd68ae1b6/panels/user-accounts/um-photo-dialog.c#L190-L192

Since gnome-desktop scans `thumbnailers` directories under the paths
in `XDG_DATA_DIRS`, gdk-pixbuf  had to be added to the path to provide
access to image thumbnailer.