CODEOWNERS: fix path notation
This format uses the path matching algorithm as the .gitignore format.

The paths need to be prefixed with a / to become absolute, otherwise it
will match all of these files.

The paths will also match any of the sub-paths by default so it's not
necessary to use the `*` like in bash globbing.